+++ packages/demo/__stories__/theming.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { storiesOf } from '@storybook/react'
 import { ThemeProvider } from 'styled-components'
 import { Story } from '.'
-import { Theming } from '@edtr-io/ui'
+import { EditorTheming } from '@edtr-io/ui'
 
 storiesOf('Theming', module).add('Initial State', () => {
   const state = {
@@ -10,7 +10,7 @@ storiesOf('Theming', module).add('Initial State', () => {
     state: [{ plugin: 'text' }]
   }
 
-  const theme: Theming = {
+  const theme: EditorTheming = {
     textColor: '#222',
     backgroundColor: '#d9edf7',
     buttonBackgroundColor: 'transparent',