+++ packages/ui/src/styled.ts
@@ -2,14 +2,14 @@ import styled from 'styled-components'
 
 export { styled }
 
-export interface Theming {
+export interface EditorTheming {
   highlightColor: string
   textColor: string
   backgroundColor: string
   buttonBackgroundColor: string
 }
 
-export const defaultTheming: Theming = {
+export const defaultTheming: EditorTheming = {
   highlightColor: 'rgb(70, 155, 255)',
   textColor: '#EEEEEE',
   backgroundColor: 'rgb(51,51,51,0.95)',