+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased](https://github.com/edtr-io/edtr-io/compare/0.4.0..HEAD)
+
+### Added
+
+- Use Theming in toolbars and menus.
+
 ## [0.4.0](https://github.com/edtr-io/edtr-io/compare/0.3.2..0.4.0) - March 21, 2019
 
 ### Breaking Changes